Some small changes and fixes.
Former-commit-id: 97ce036662a5a9ca31a4d7f900d5e1cf0933e3f5
Former-commit-id: 111395a04b3e6e66e2a8d78b82e0ee4b38db85a8

+++ App/src/canvas.cpp
@@ -611,7 +611,7 @@ void Canvas::drawCross()
 
 	double position = (getInfoTable()->getPosition() + pos.x())*ratio;
 
-	float data[8] = {static_cast<float>(position), 0, static_cast<float>(position), static_cast<float>(height()), 0, pos.y(), static_cast<float>(samplesRecorded), pos.y()};
+	float data[8] = {static_cast<float>(position), 0, static_cast<float>(position), static_cast<float>(height()), 0, static_cast<float>(pos.y()), static_cast<float>(samplesRecorded), static_cast<float>(pos.y())};
 
 	gl()->glBufferData(GL_ARRAY_BUFFER, sizeof(data), data, GL_STATIC_DRAW);
